- The method for specifying AI and revisions for multiple instances has changed. You now enter them in a single line. Example: someAI.dll:3400, secondInstance.dll, thirdInstance.dll:3650, fourth.dll - The config file now loads a DLL from "ai" and "ai_dbg" instead of "ai_dll" to maintain compatibility with older revisions. - Removed RTree. Now using Starcraft's native unit searches which are more efficient. - Currently getUnitsInWeaponRange is slightly broken with an unknown problem (I will try to look into it with more detail). - Made several getDistance functions return an integer instead of double (they were returning an integer cast to a double, so this should increase performance). - Added the ability to change the screenshot format via config ("screenshots" under "Starcraft"). It is set to GIF by default. - Added Unit::getUnitsInRadius for spellcasting and other uses.
git-svn-id: https://bwapi.googlecode.com/svn/trunk@3466 273ecf2c-f24b-0410-b6a7-e9298bb4873f

+++ bwapi/BWAPIClient/Source/GameImpl.cpp
@@ -494,7 +494,7 @@ namespace BWAPI
     return empty;
   }
   //----------------------------------------------- GET UNITS IN RADIUS --------------------------------------
-  std::set<Unit*> GameImpl::getUnitsInRadius(BWAPI::Position center, double radius)
+  std::set<Unit*> GameImpl::getUnitsInRadius(BWAPI::Position center, int radius)
   {
     //TODO: Implement R-Tree on BWAPI Client (and add a way to disable it on server)
     std::set<Unit*> empty;
